Добавил удаление и подсчет вложенных комментариев

+++ packages/server/controllers/comment.ts
@@ -3,6 +3,7 @@ import { User } from '../models/user';
 import { Comment } from '../models/comment';
 import type { IComment } from 'comment';
 import { errorMessage } from '../utils/messageHelper';
+import { Op } from 'sequelize';
 
 /**
  * Пример запроса
@@ -12,7 +13,27 @@ import { errorMessage } from '../utils/messageHelper';
  */
 export const commentCreate = async (req: Request, res: Response) => {
   try {
+    const { comment_id } = req.body;
     const comment = await Comment.create(req.body);
+
+    /** Если у созданного комментария есть родительский коммент,
+     * то увеличеваем у нгео счетчик вложенных комментариев */
+    if (comment_id) {
+      const findParentComment = await Comment.findByPk(comment_id);
+
+      if (!findParentComment) {
+        return res.status(404).json(errorMessage(`Комментария с id ${comment_id} не найдено`))
+      }
+
+      const updateObj = {
+        nested_comment_count: findParentComment?.nested_comment_count + 1
+      }
+
+      await Comment.update(updateObj, {
+        where: { id: comment_id }
+      });
+    }
+
     return res.status(200).json(comment.dataValues);
   } catch (error) {
     return res.status(500).json({ message: 'error', error: error })
@@ -24,7 +45,10 @@ export const commentGet = async (req: Request<{ topic_id: number }>, res: Respon
     const { topic_id } = req.params;
     const comments = await Comment.findAll({
       where: {
-        topic_id
+        topic_id,
+        comment_id: {
+          [Op.not]: null
+        }
       }
     })
 
@@ -99,15 +123,48 @@ export const commentUpdate = async (req: Request, res: Response) => {
 export const commentDelete = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
-    const comment = await Comment.destroy({
-      where: { id }
-    });
+    const comment = await Comment.findByPk(id);
+
+    /** Если у удаленного комментария есть родительский коммент,
+     * то уменьшаем у нгео счетчик вложенных комментариев */
+    if (comment?.comment_id) {
+      const findParentComment = await Comment.findByPk(comment?.comment_id);
+
+      if (!findParentComment) {
+        return res.status(404).json(errorMessage(`Комментария с id ${comment?.comment_id} не найдено`))
+      }
+
+      const updateObj = {
+        nested_comment_count: findParentComment?.nested_comment_count - 1
+      }
+
+      await Comment.update(updateObj, {
+        where: { id: comment?.comment_id }
+      });
+    }
+
+    if (!comment) {
+      return res.status(404).json(errorMessage('Comment not found'))
+    }
+
+    if (comment.comment_id === null) {
+      await Comment.destroy({
+        where: {
+          comment_id: comment.id,
+        }
+      });
+      await comment.destroy();
+    } else {
+      await comment.destroy();
+    }
+
+
 
     if (comment) {
       return res.status(204).json({ message: 'Comment deleted' })
     }
 
-    return res.status(404).json(errorMessage('User not found'))
+    return res.status(404).json(errorMessage('Comment not found'))
   } catch (error) {
     return res.status(500).json(errorMessage(error))
   }
